[android]: Merge two StorageUtils classes into one
No semantic changes intended. This patch is pure refactoring.
