+++ android/src/com/mapswithme/maps/settings/StoragePathManager.java
@@ -22,6 +22,7 @@ import com.mapswithme.maps.R;
 import com.mapswithme.maps.dialog.DialogUtils;
 import com.mapswithme.maps.downloader.MapManager;
 import com.mapswithme.util.Config;
+import com.mapswithme.util.StorageUtils;
 import com.mapswithme.util.concurrency.ThreadPool;
 import com.mapswithme.util.concurrency.UiThread;
 import com.mapswithme.util.log.Logger;
